+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Qasir React Boilerplate
 
+**Coming Soon:**
+- Adding hot module reloading
 
-**Release Note:**
+**Release Note:**    
 
 - V1.0.4 (2019-09-05)    
     - Update server script not support redis server
@@ -25,9 +27,9 @@
     - Adding views folder for handle website pages
     - Update webpack configuration
     - Update ignore files    
-    - Update package position
-    - Adding hot module reloading
+    - Update package position    
     - Adding ESLint code linting (Airbnb linter)
     - Update routing system
     - Adding PWA config support (workbox)
     - Adding manifest file
+